Phase 6 scoping + commit 6a: fleet/starbase movement fidelity
docs/ROADMAP.md and PORT_DESIGN.md: six research forks read every
NPE-adjacent Pascal file (~4,400 lines) before planning. Real Pascal has
four structurally distinct NPE personalities (Pirate/Kingdom/Berserker/
Guardian) behind one variant-typed dispatcher, not "one AI, two tiers" as
the roadmap's original framing assumed. Checked reachability against all
12 dos_131 scenarios: Kingdom dominates every scenario with any NPE at
all; Guardian, the smallest by line count, never actually appears.
TraderNPE is confirmed dead (zero dispatch arms, zero scenario usage).
This phase builds Kingdom plus shared infrastructure sized for the other
personalities later, per user direction. Full design rationale in
PORT_DESIGN.md; commit breakdown (6a-6f) in ROADMAP.md.

6a itself: FleetMovementHandler's fuel model was invented, not ported --
compared against DATACNST.PAS's real FuelCons/FuelCap tables while
scoping and found it backwards in places (e.g. Jumpships cheaper than
Starships, the reverse of source). Replaced with FleetLogistics.cs
(FuelCapacity/FuelConsumption/FleetCargoSpace/BalanceFleet, MISC.PAS
verbatim) and made Fleet.Fuel a double to match Pascal's own Real-typed
fuel arithmetic. Dense-nebula blocking is now a real per-step check for
every fleet type, not just Jump/HunterKiller. Stargate teleportation and
fortress pass-through jumps are real (FLEET.PAS's UpdateFleet) --
including the non-obvious case where a fortress hop that lands short of
teleport range still gets its ordinary per-turn movement allotment on
top, same turn. UpdateAllFleets' dispatch now checks gate occupancy the
way source does (owner-blind), via new Galaxy.GetObjectAt. Starbase
movement (MovePlayerStarbases/GetNewBasePos) is now obstacle-avoiding
and fuel-costed instead of a straight-line stepper -- removed
YearsUntilNextMove, an invented throttle with no source counterpart
(confirmed via MovePlayerStarbases' actual call sites).

FLEET.PAS's ground-truth patch (reference/verify/) is still pending --
next up.

+++ src/ThreeLn.Reconstruction4021.Core/Entities/Fleet.cs
@@ -13,9 +13,14 @@ public sealed class Fleet : IMovable, ISectorObject
 
     /// <summary>
     /// Pascal splits this across FuelHigh/Fuel to dodge 16-bit Integer overflow
-    /// (FuelHigh*MaxInt + Fuel) — no reason to carry that split into C#.
+    /// (FuelHigh*MaxInt + Fuel) — no reason to carry that split into C#. Real (not int): Pascal's own
+    /// GetFleetFuel/SetFleetFuel expose it as a Real everywhere it's actually used
+    /// (UseUpFuel/FuelConsumption/FuelCapacity all compute fractionally, truncating only when the
+    /// value is written back to the underlying storage) — an int here would round a sub-1.0 yearly
+    /// consumption (a lone fighter burns ~0.01/year, MISC.PAS:394-396) to 0 or 1 every turn instead of
+    /// accumulating fractionally, a different mechanic, not just lost precision.
     /// </summary>
-    public int Fuel { get; set; }
+    public double Fuel { get; set; }
 
     public Coordinate? Destination { get; set; }
     public FleetStatus Status { get; set; } = FleetStatus.Ready;
